- break up the lazyloader into individual files
