+++ src/components/DatePickerRange.react.js
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types';
 import React, {Component, lazy, Suspense} from 'react';
-import {datePickerRange} from '../utils/LazyLoader';
+import datePickerRange from '../utils/LazyLoader/datePickerRange';
 
 const RealDatePickerRange = lazy(datePickerRange);
