+++ src/utils/LazyLoader/datePickerRange.js
@@ -1,0 +1,4 @@
+export default () =>
+    import(/* webpackChunkName: "datepicker" */ '../../fragments/DatePickerRange.react');
+
+
