+++ src/utils/LazyLoader/datePickerSingle.js
@@ -1,0 +1,2 @@
+export default () => import(/* webpackChunkName: "datepicker" */ '../../fragments/DatePickerSingle.react');
+
