+++ src/utils/LazyLoader/dropdown.js
@@ -1,0 +1,2 @@
+export default () => import(/* webpackChunkName: "dropdown" */ '../../fragments/Dropdown.react');
+
