+++ src/utils/LazyLoader/graph.js
@@ -1,0 +1,2 @@
+export default () => import(/* webpackChunkName: "graph" */ '../../fragments/Graph.react');
+
