+++ src/utils/LazyLoader/markdown.js
@@ -1,0 +1,2 @@
+export default () => import(/* webpackChunkName: "markdown" */ '../../fragments/Markdown.react');
+
