+++ src/utils/LazyLoader/plotly.js
@@ -1,0 +1,6 @@
+export default () => Promise.resolve(window.Plotly ||
+    import(/* webpackChunkName: "plotlyjs" */ 'plotly.js').then(({ default: Plotly }) => {
+        window.Plotly = Plotly;
+        return Plotly;
+    }));
+
