+++ src/utils/LazyLoader/upload.js
@@ -1,0 +1,1 @@
+export default () => import(/* webpackChunkName: "upload" */ '../../fragments/Upload.react');
